feat: Implement update assessment fucnitonality
- Create a controller that handle assessment update

- Create a new service method that handles deleting of upladed files when assessment was updated

- Create a route and controller for unpublishing assessment

+++ database/migrations/2025_08_14_044130_create_assessments_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('assessment_title');
             $table->longText('assessment_description')->nullable();
             $table->dateTime('due_datetime')->nullable();
-            $table->decimal('total_points');
+            $table->integer('total_points');
             $table->uuid('created_by');
             $table->enum('status', ['published', 'draft']);
             $table->foreign('created_by')->references('user_id')->on('users');
